test/implements/UserDB.test.ts
const chai = require("chai");
const expect = chai.expect;
import UserDb from "../../src/implements/UserDB";
import { resGen } from "../../src/models/resGen";

describe("testing clase UserDb.", function () {
  it("Test of function excuteAddUserDb.", async function () {
    let myUserDb = new UserDb();
    let myResGen = new resGen();
    let user = { user: "userTest", email: "email@gmail.com" };
    myResGen = await myUserDb.DB_AddUser(user);
    myResGen.valid = true;
    myResGen.message = "OPERATION_SUCCESFULL";
    expect(myResGen.valid).to.equal(true);
    expect(myResGen.message).to.equal("OPERATION_SUCCESFULL");
    expect(myResGen.message).to.be.a("string");
  });
});
